+++ packages/hooks/use-registry/README.md
@@ -1,0 +1,11 @@
+# `@hi-ui/use-registry`
+
+> TODO: description
+
+## Usage
+
+```
+const useRegistry = require('@hi-ui/use-registry');
+
+// TODO: DEMONSTRATE API
+```
