+++ packages/hooks/use-registry/__tests__/use-registry.test.js
@@ -1,0 +1,5 @@
+const useRegistry = require('../src')
+
+describe('@hi-ui/use-registry', () => {
+  it('needs tests', () => {})
+})
