+++ packages/hooks/use-registry/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../../../jest.config')
